ci: exclude git directory in deployment workflow
Exclude `.git` directory in continuous deployment workflow. This reduces
the deployment size, allows for faster deployments, better for security
as it could exposes unnecessary details to our project's history, and
there is simply no practical use for the folder in production; it
doesn't need version control history, it simply needs to run the
application.